// plugins/plugin-harbor/src/utils/interfaces.ts
export interface IOptions {
  query: string;
  page: number;
  perPage: number;
}
// ...
export interface IRepositoriesData {
  repositories: IRepository[];
  total: number;
}

export interface IRepository {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  artifact_count: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  creation_time: string;
  id: number;
  name: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  project_id: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  pull_count: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  update_time: string;
}